% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meta_info_from_labels}
\alias{meta_info_from_labels}
\title{Generate meta-information from sample names}
\usage{
meta_info_from_labels(sample_names)
}
\arguments{
\item{sample_names}{A character vector of sample names}
}
\value{
A tibble containing sample_name, timepoint, and replicate columns
}
\description{
Generate meta-information from sample names
}
