% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_replicate}
\alias{sample_replicate}
\title{Extract replicate from sample name}
\usage{
sample_replicate(x)
}
\arguments{
\item{x}{A character vector of sample names}
}
\value{
A character vector of replicates
}
\description{
Extract replicate from sample name
}
