% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timepoint_from_sample}
\alias{timepoint_from_sample}
\title{Extract timepoint from sample name}
\usage{
timepoint_from_sample(x)
}
\arguments{
\item{x}{A character vector of sample names}
}
\value{
A character vector of timepoints
}
\description{
Extract timepoint from sample name
}
